requirements.txt
# 웹 프레임워크
Flask==3.0.0
Werkzeug==3.0.1

# 데이터베이스
psycopg2-binary==2.9.9
SQLAlchemy==2.0.23

# 엑셀 파일 처리
pandas==2.1.4
openpyxl==3.1.2

# 환경 변수 관리
python-dotenv==1.0.0

# 웹 서버 (배포용)
gunicorn==21.2.0
